+++ slam/buffer.h
@@ -15,7 +15,7 @@ int buffer_get_width();
 int buffer_get_height();
 uint8_t* buffer_allocate();
 void buffer_attenuate(uint8_t*, double);
-int buffer_index_from_x_y(double, double);
+int buffer_index_from_x_y(int, int);
 int x_from_buffer_index(int);
 int y_from_buffer_index(int);
 int index_protected(int);
